pyproject.toml
[tool.poetry]
name = "tiktok-downloader-tgbot"
version = "0.1.0"
description = ""
authors = ["Vyacheslav1557 <slavabirin1@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^22.0"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
yt-dlp = "^2025.4.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
